fix(rpcclient): serve WebDAV Range requests to prevent large-file corruption
The WebDAV provider declared content_range=False and ignored the HTTP Range
header, answering every GET with 200 and the entire file body. macOS Finder /
webdavfs reads large files as a series of byte ranges; receiving the whole file
in response to `Range: bytes=X-Y` makes the client write the full body at offset
X, producing a corrupted file.

Honor Range requests like asgi_webdav's own FileSystemProvider: advertise
Accept-Ranges, return 206 Partial Content with a body generator that seeks to
the requested offset and streams only the requested bytes (with 416 handling for
a failed If-Range). Also widen File.seek to c_int64 so range offsets past 2 GiB
are not truncated.

+++ src/rpcclient/rpcclient/core/subsystems/fs.py
@@ -148,7 +148,7 @@ class File(Allocated[ClientT_co]):
 
     async def seek(self, offset: int, whence: int) -> int:
         """Seek the remote file descriptor and return the resulting offset."""
-        err = (await self._client.symbols.lseek(self.fd, offset, whence)).c_int32
+        err = (await self._client.symbols.lseek(self.fd, offset, whence)).c_int64
         if err < 0:
             await self._client.raise_errno_exception(f"failed to lseek fd: {self.fd}")
         return err
